Fixed .ply io inconsistency, added coneZ function
Files written a ply format were always in binary, and read in ascii,
so, a file written by this package could not be read by this package.
Now this has been clarified in the function calls.  Modified README.md
accordingly, and added a test.
Added a new csg function for a cone aligned with z axis.